pyproject.toml
[tool.poetry]
name = "homelab-playbooks"
version = "0.1.0"
description = ""
authors = ["Bakie <bakelants.developer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
ansible = "^4.8.0"
"github3.py" = "^3.0.0"


[tool.poetry.dev-dependencies]
ansible-lint = "^5.2.1"
yamllint = "^1.26.3"
flake8 = "^4.0.1"
molecule = "^3.5.2"
molecule-docker = "^1.1.0"
docker = "^5.0.3"
pytest-testinfra = "^6.4.0"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
